Return int32 instead of double from wall time. Add tests for afro esc. Use crocodile for all includes (added to path).

+++ drone/inc/wall_time.h
@@ -2,12 +2,13 @@
 #define WALL_TIME_H
 
 #include <chrono>
+#include <cstdint>
 
 class WallTime
 {
 public:
-    double millis();
-    double micros();
+    uint32_t millis();
+    uint32_t micros();
 };
 
 extern WallTime wall_time;
